Promote experimental layout management to non-experimental
