+++ app/components/LayoutStorageReduxAdapter.tsx
@@ -4,12 +4,10 @@
 
 import { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { useAsync } from "react-use";
-import { useDebounce } from "use-debounce";
+import { useAsync, useThrottle } from "react-use";
 import { v4 as uuidv4 } from "uuid";
 
 import { loadLayout } from "@foxglove-studio/app/actions/panels";
-import { useExperimentalFeature } from "@foxglove-studio/app/context/ExperimentalFeaturesContext";
 import { useLayoutStorage } from "@foxglove-studio/app/context/LayoutStorageContext";
 import { State } from "@foxglove-studio/app/reducers";
 import sendNotification from "@foxglove-studio/app/util/sendNotification";
@@ -20,43 +18,34 @@ export default function LayoutStorageReduxAdapter() {
   const dispatch = useDispatch();
 
   // Debounce the panel state to avoid persisting the layout constantly as the user is adjusting it
-  const [debouncedState] = useDebounce(panelsState, 1000 /* 1 second */);
+  const throttledPanelsState = useThrottle(panelsState, 1000 /* 1 second */);
 
   const layoutStorage = useLayoutStorage();
-  const enableLayoutManagement = useExperimentalFeature("layoutManagement");
 
   // save panel state to our storage
   const { error } = useAsync(async () => {
-    if (!enableLayoutManagement) {
-      return;
-    }
-
-    if (debouncedState.id === undefined) {
+    if (throttledPanelsState.id === undefined) {
       return;
     }
 
     const layout = {
-      id: debouncedState.id,
-      name: debouncedState.name ?? "unnamed",
-      state: debouncedState,
+      id: throttledPanelsState.id,
+      name: throttledPanelsState.name ?? "unnamed",
+      state: throttledPanelsState,
     };
 
     // save to our storage
     await layoutStorage.put(layout);
-  }, [enableLayoutManagement, layoutStorage, debouncedState]);
+  }, [layoutStorage, throttledPanelsState]);
 
   // set an id if panel is missing one
   useEffect(() => {
-    if (!enableLayoutManagement) {
-      return;
-    }
-
     if (panelsState.id === undefined) {
       panelsState.id = uuidv4();
       panelsState.name = panelsState.name ?? "unnamed";
       dispatch(loadLayout(panelsState));
     }
-  }, [dispatch, enableLayoutManagement, panelsState]);
+  }, [dispatch, panelsState]);
 
   useEffect(() => {
     if (error) {
