+++ app/experimentalFeatures.tsx
@@ -38,10 +38,4 @@ export default {
     developmentDefault: false,
     productionDefault: false,
   },
-  layoutManagement: {
-    name: "Manage Layouts",
-    description: `Manage layouts via the layout menu. Add/Remove/Copy layouts. Import/Export layouts to files.`,
-    developmentDefault: true,
-    productionDefault: true,
-  },
 } as FeatureDescriptions;
